+++ .config/webpack/webpack.config.ts
@@ -197,14 +197,14 @@ const config = async (env): Promise<Configuration> => {
       modules: [path.resolve(process.cwd(), 'src'), 'node_modules'],
       unsafeCache: true,
     },
-  };
+  }
 
-  if (isWSL()) {
+  if(isWSL()) {
     baseConfig.watchOptions = {
       poll: 3000,
       ignored: /node_modules/,
-    };
-  }
+    }}
+
 
   return baseConfig;